% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFGaussBounds.R
\name{QFGaussBoundsNew}
\alias{QFGaussBoundsNew}
\title{New Bounds on the CDF of a Quadratic Form in Gaussians}
\usage{
QFGaussBoundsNew(cdf, f = "identity", max.abs.eta, sum.eta, sum.etasq,
  sum.eta.deltasq = 0, sum.etasq.deltasq = 0)
}
\description{
New Bounds on the CDF of a Quadratic Form in Gaussians
}
