% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFIntBounds2.R
\name{QFIntBounds2}
\alias{QFIntBounds2}
\title{Quadratic form bounds}
\usage{
QFIntBounds2(obs, evals, ncps, E_R, nu, resid.op.norm.bd, lower.tail = TRUE,
  log = FALSE, llin.approx = FALSE)
}
\arguments{
\item{obs}{scalar; the observed value of the quadratic form: that is, \eqn{y^T M y}.}

\item{evals}{vector of eigen-values of the matrix \eqn{M}.  These need not be all eigen values.}

\item{ncps}{description here}

\item{E_R}{description here}

\item{nu}{description here}

\item{resid.op.norm.bd}{description here}

\item{lower.tail}{logical; if \code{TRUE} (default), probability is \eqn{P(y^T M y \le obs)}, otherwise \eqn{P(y^T M y > obs)}}

\item{log}{logical; if \code{TRUE}, probability \eqn{p} is given as \eqn{log(p)}}

\item{N}{description here}

\item{if.insuff.eigs}{string indicating what action to take if there are insufficient eigen-values to produce an accurate bound.  If \code{"trivial"} then the bounds \eqn{[0,1]} are returned; if \code{"missing"} then \code{NA} is returned for both bounds.}
}
\value{
A data frame containing the variables lower and upper which provide the bounds on the CDF of the quadratic form.
}
\description{
Compute bounds on the CDF of a quadratic form.
}
\details{
The above was the one line summary which appears at the top of the documentation.  The gory detailed description of what is does goes here.

Those gory details can span as many paragraphs as you want.
}
\examples{
# Some code here which runs a self-contained example

}
