% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFGauss.R
\name{TestQFGauss}
\alias{TestQFGauss}
\title{Test function for a QFGaussCDF object}
\usage{
TestQFGauss(cdf, n.samps = 10000)
}
\arguments{
\item{cdf}{a QFGaussCDF}

\item{n.samps}{number of draws from the target distribution with which to construct the empirical CDF}
}
\value{
Nothing is returned.
}
\description{
Compares the CDF inferred by QFGauss to an empirical CDF.
}
\details{
Four plots are produced.  The top-left plot overlays the CDF computed by QFGauss (in black) and the empirical CDF (in red) based on 10,000 samples.
The top-right plot shows the distance between the empirical and QFGauss-computed CDF and the corresponding ks.test p-value (two-sided alternative).
The ks.test p-value will be \code{NA} if \code{cdf} is missing one of its tails (see \code{\link{QFGauss}} for details).
The two bottom plots allow comparison of the empirical CDF (in red) with the computed CDF (in black) in each tail.
}
\examples{
TestQFGauss(QFGauss(c(1,3,4,-3)))

}
\seealso{
\code{\link{QFGauss}}, \code{\link{TestQFGaussBounds}}
}
