% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFGaussBounds.R
\name{TestQFGaussBounds}
\alias{TestQFGaussBounds}
\title{Test function for a QFGaussBounds}
\usage{
TestQFGaussBounds(fullcdf, k = min(20, floor(length(attr(fullcdf,
  "f.eta"))/2)), n.bound.points = 20)
}
\arguments{
\item{fullcdf}{a QFGaussCDF of the target distribution including all terms}

\item{k}{the number of truncated terms provided to QFGaussBounds from which to bound fullcdf}

\item{n.bound.points}{the number of points at which to evaluate the bounds for plotting}
}
\value{
There is nothing returned.
}
\description{
Compares the CDF bounds inferred by QFGaussBounds to a truncated approximation of the CDF and a naive quadrature-based implementation of the bounds.
}
\details{
Four plots are produced.  The top-left plot overlays the target, \code{fullcdf}, computed by \code{QFGauss} (in black) and a truncated approximation to that CDF (in orange) based on simply adding the expectation of the remainder terms to the top-k truncated version of \code{fullcdf}.
The upper and lower bounds on \code{fullcdf} computed by \code{QFGaussBounds} are plotted as red and blue circles respectively.
The upper and lower bounds on \code{fullcdf} computed by a naive quadrature-based implementation of the bounds are plotted as red and blue Xs respectively.
The top-right plot shows the difference between the truncated approximation of the CDF and \code{fullcdf} in log space.  It may be interpretted as follows.  The x-axis plots the -log_10 p-value one would have reported based on the truncated approximation alone.  The y-axis is the difference between
the true -log_10 p-value and the approximate -log_10 p-value.  The difference in p-values in the upper tail is plotted with a solid line.  The difference in p-values in the lower tail is plotted with a dashed line.  This plot effectively shows how far one might be misled by the truncated approximation shifted by the expectation of the remainder terms.
The two bottom plots show the top-left plot, just on the log-scale, to focus on each tail separately.
}
