% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFGauss.R
\name{plot.QFGaussCDF}
\alias{plot.QFGaussCDF}
\title{Plot method for a QFGaussCDF object}
\usage{
\method{plot}{QFGaussCDF}(x, ...)
}
\arguments{
\item{...}{additional parameters}

\item{cdf}{a QFGaussCDF}
}
\value{
There is nothing returned.
}
\description{
Plots the CDF computed by QFGauss.
}
