% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QForm.R
\name{QForm}
\alias{QForm}
\title{Quadratic Form Bounds}
\usage{
QForm(lambda, delta = rep(0, length(lambda)), n = 2^16 - 1)
}
\arguments{
\item{obs}{vector; observed values of the quadratic form for which upper and lower bounds on the CDF}

\item{qf.cdf}{QForm cdf object; output of QFcdf for some set of coefficients and ncps}

\item{fun}{character string; function to be applied to coefficients.  Current options: "identity","power","exponential"}

\item{fun.args}{list; list of function arguments required for the corresponding function type provided in fun}

\item{lower.tail}{logical; if \code{TRUE} (default), probability is \eqn{P(y^T M y \le obs)}, otherwise \eqn{P(y^T M y > obs)}}

\item{log}{logical; if \code{TRUE}, probability \eqn{p} is given as \eqn{log(p)}}
}
\value{
A data frame containing the variables lower and upper which provide the bounds on the CDF of the quadratic form.
}
\description{
Compute upper and lower bounds on the CDF of a quadratic form in normal random variables.
}
\details{
Gory details....
}
\examples{
# Some code here which runs a self-contained example

}
