% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFGauss.R
\name{TestQFGauss}
\alias{TestQFGauss}
\title{Test function for a QFGaussCDF object}
\usage{
TestQFGauss(cdf, n.samps = 10000)
}
\arguments{
\item{cdf}{a QFGaussCDF}

\item{n.samps}{number of draws from the target distribution with which to construct the empirical CDF}
}
\value{
There is nothing returned.
}
\description{
Compares the CDF inferred by QFGauss to an empicical CDF.
}
\details{
Four plots are produced.  The top-left plot overlays the CDF computed by QFGauss (in black) and the empirical CDF (in red) based on 10,000 samples.
The top-right plot shows the distance between the empirical and QFGauss-computed CDF and the corresponding ks.test p-value (two-sided alternative).
The two bottom plots compare the empirical CDF (in red) with the computed CDF (in black) in each tail.
}
