% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFGaussBounds.R
\name{TestQFGaussBounds}
\alias{TestQFGaussBounds}
\title{Test function for a QFGaussBounds}
\usage{
TestQFGaussBounds(
  fullcdf,
  k = min(20, floor(length(attr(fullcdf, "f.eta"))/2)),
  n.bound.points = 16,
  lower.tail.end = 20,
  upper.tail.end = 20,
  parallel.sapply = base::sapply
)
}
\arguments{
\item{fullcdf}{QFGaussCDF; the target CDF including all terms; currently TestQFGaussBounds only works for cases where the QFGauss produeced CDF has all df = 1.}

\item{k}{numeric; the number of truncated terms provided to QFGaussBounds from which to bound fullcdf}

\item{n.bound.points}{numeric; the number of points at which to evaluate the bounds for plotting}

\item{lower.tail.end}{numeric; the -log_10 lower tail probability at which to start each x-axis (default = 20)}

\item{upper.tail.end}{numeric; the -log_10 upper tail probability at which to end each x-axis (default = 20 )}

\item{parallel.sapply}{function; a user-provided version of \code{sapply}, see Details.}
}
\value{
There is nothing returned.
}
\description{
Compares the CDF bounds inferred by QFGaussBounds to a truncated approximation of the CDF and a naive quadrature-based implementation of the bounds.
}
\details{
Here, \code{fullcdf} is taken to be the CDF of the target random variable \eqn{Q_f} (see documentation of \code{\link{QFGauss}} for definitions).
Four plots are produced.  The top-left plot overlays the target CDF of \eqn{Q_f}, \code{fullcdf}, computed by \code{QFGauss} (in black) and a truncated approximation to that CDF (in orange) based on simply adding the expectation of the remainder term \eqn{R_f} to the top-k truncated version of \code{fullcdf}, \eqn{T_f}.  By "top-k" here we mean taking the terms of \eqn{Q_f} with the largest magnitude coefficients, \code{f.eta}, and using that to define \eqn{T_f}, which is what is done in \code{TestQFGaussBounds} internally.
The green line is a similar approximation but where \eqn{R_f} is approximated with a moment-matching gaussian.
The upper and lower bounds on \code{fullcdf} computed by \code{QFGaussBounds} are plotted as red and blue circles respectively.
The upper and lower bounds on \code{fullcdf} computed by a naive quadrature-based implementation of the bounds are plotted as red and blue Xs respectively.
The top-right plot shows the difference between the truncated approximation of the CDF and \code{fullcdf} in log space.  It may be interpreted as follows.  The x-axis plots the -log_10 p-value one would have reported based on the truncated approximation alone.  The y-axis is the difference between
the true -log_10 p-value and the approximate -log_10 p-value.  The difference in p-values in the upper tail is plotted with a solid line.  The difference in p-values in the lower tail is plotted with a dashed line.  This plot effectively shows how far one might be misled by the truncated approximation shifted by the expectation of the remainder terms.
The two bottom plots allow comparison of the empirical CDF (in red) with the computed CDF (in black) in each tail.
}
\examples{
TestQFGaussBounds(QFGauss(c(1,5,-4,-3),c(2,-1,4,-5)^2),2)
}
\seealso{
\code{\link{QFGaussBounds}}, \code{\link{TestQFGauss}}
}
